pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.92",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "mangum>=0.19.0",
    "pydantic-settings>=2.7.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.174.0",
    "constructs>=10.4.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.1",
    "pytest-mock>=3.14.0",
]
